# api/openapi3_0.yaml
openapi: 3.0.3
info:
  title: Game Score Manager API
  version: 1.3.9
  description: Swagger for Game Score manager
  contact:
    email: desarrollo070615@gmail.com
    name: Leidy Santos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3000/api/v1/
  - url: http://172.10.10.1:3000/api/v1/
...
paths:
...
  /users/admin:
   get:
      summary: Get all users
      security:
        - bearerAuth: []
      tags:
        - Users
      description: Get all users by admin
      parameters:
        - name: page
          in: query
          description: Actual page
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of elements per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: All users uploaded successfully
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UserPagination'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse401'
        '400':
          description: User bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse400'
        '403':
          description: Valid token, no permission for this action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse403'        
  /users/admin/{userId}:
    patch:
      summary: Enable or block user
      security:
        - bearerAuth: []
      tags:
        - Users
      description: Enable or block user by admin
      parameters:
        - name: userId
          in: path
          description: The ID that need to be fetched.
          required: true
          schema: 
            type: string 
      responses:
        '200':
          description: User enabled or block successfully
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse401'
        '400':
          description: User bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse400'
        '403':
          description: Valid token, no permission for this action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse403'        
    delete:
      summary: Delete user
      security:
        - bearerAuth: []
      tags:
        - Users
      description: Delete user by admin
      parameters:
        - name: userId
          in: path
          description: The ID that need to be fetched.
          required: true
          schema: 
            type: string 
      responses:
        '200':
          description: User removed successfully
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteUserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse401'
        '400':
          description: User bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse400'
        '403':
          description: Valid token, no permission for this action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse403'        
...
components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
          example: 45247989-8807-4cec-bf65-50632859a8f5
        name:
          type: string
          example: Leidy Santos
        username:
          type: string
          example: lsantos
        email:
          type: string
          format: email
          example: prueba@yopmail.com
        role:
          type: string
          example: admin
        token: 
          type: string
          example: HHOl6ie7sb4NBYmypXOKmo5cVB8IMy7gm6xDqI4XeDthmS5GPSI1oabdxIShX8r1
        status:
          type: string
          example: Valid
    UserPagination:
      type: object
      properties:
        currentPage:
          type: integer
          description: actual page
          example: 1
        totalPages:
          type: integer
          description: Total available pages
          example: 10
        totalUsers:
          type: integer
          description: Total users
          example: 20
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
    DeleteUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuId
          example: 45247989-8807-4cec-bf65-50632859a8f5
        name:
          type: string
          example: Leidy Santos
        username:
          type: string
          example: lsantos
        email:
          type: string
          format: email
          example: prueba@yopmail.com
        password:
          type: string
          format: password
          example: 114c884d2a1cd1cde577ff109c3db5c6
        role:
          type: string
          example: admin
        token:
          type: string
          example: HHOl6ie7sb4NBYmypXOKmo5cVB8IMy7gm6xDqI4XeDthmS5GPSI1obdxIShX8r1
        status:
          type: string
          example: Invalid      
...
    ErrorResponse400:
      type: object
      properties:
        message:
          type: string
          example: Invalid data
        statusCode:
          type: integer
          example: 400
...
    ErrorResponse401:
      type: object
      properties:
        message:
          type: string
          example: Unauthorized
        statusCode:
          type: integer
          example: 401
    ErrorResponse403:
      type: object
      properties:
        message:
          type: string
          example: Valid token, but you don't have access to this functionality
        statusCode:
          type: integer
          example: 403
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT        